phase 2
Looks great, don't think I'll have to change to many things to make it
work as is. but I will change some things so that is less restrictive,
like forcing people to use a command for localized chat is retarded. But
forcing them to use a command for server wide now that is on the mark
and should be done.