phase 5?
Everything I wanted to keep is ported up now. the red squiggles are
gone. Now I have to set a server environment to test it. Well I should
try to run it first and see where it will crash first.

And there are some tweaks that have to be made so the default talk
command is the same thing that happens when type on the chat line
without a /